k8s: get CNP from apiserver if nil in `updateStatusWithCapabilities`
Future commits will introduce cases where the CNP may not be available from
callers of this function. If that is the case, it will be nil. Provide logic to
try to get the CNP via the client to the K8s apiserver in such cases.

Signed-off by: Ian Vernon <ian@cilium.io>